+++ api/plan-dft-1d.c
@@ -21,7 +21,7 @@
 #include "api.h"
 #include "dft.h"
 
-X(plan) X(plan_dft_1d)(unsigned int n, C *in, C *out, int sign, 
+X(plan) X(plan_dft_1d)(unsigned long n, C *in, C *out, int sign, 
 		       unsigned int flags)
 {
      return X(plan_dft)(1, &n, in, out, sign, flags);
