+++ api/plan-dft-r2c-3d.c
@@ -20,10 +20,10 @@
 
 #include "api.h"
 
-X(plan) X(plan_dft_r2c_3d)(unsigned int nx, unsigned int ny, unsigned int nz,
+X(plan) X(plan_dft_r2c_3d)(unsigned long nx, unsigned long ny, unsigned long nz,
 			   R *in, C *out, unsigned int flags)
 {
-     unsigned int n[3];
+     unsigned long n[3];
      n[0] = nx; n[1] = ny; n[2] = nz;
      return X(plan_dft_r2c)(3, n, in, out, flags);
 }
