+++ api/plan-dft-r2c.c
@@ -20,7 +20,7 @@
 
 #include "api.h"
 
-X(plan) X(plan_dft_r2c)(unsigned int rank, const unsigned int *n,
+X(plan) X(plan_dft_r2c)(unsigned int rank, const unsigned long *n,
 			R *in, C *out, unsigned int flags)
 {
      return X(plan_many_dft_r2c)(rank, n, 1,
