+++ api/plan-many-dft-c2r.c
@@ -21,16 +21,16 @@
 #include "api.h"
 #include "rdft.h"
 
-X(plan) X(plan_many_dft_c2r)(unsigned int rank, const unsigned int *n,
-			     unsigned int howmany,
-			     C *in, const unsigned int *inembed,
-			     int istride, int idist,
-			     R *out, const unsigned int *onembed,
-			     int ostride, int odist,
+X(plan) X(plan_many_dft_c2r)(unsigned int rank, const unsigned long *n,
+			     unsigned long howmany,
+			     C *in, const unsigned long *inembed,
+			     long istride, long idist,
+			     R *out, const unsigned long *onembed,
+			     long ostride, long odist,
 			     unsigned int flags)
 {
      R *ri, *ii;
-     uint *nfi, *nfo;
+     ulong *nfi, *nfo;
      int inplace;
      X(plan) p;
